pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bulk-webp-url-replacer"
version = "0.1.0"
description = "Bulk convert images to WebP and automatically update URLs in markdown files"
readme = "README.md"
license = "MIT"
authors = [
    { name = "HoangYell", email = "hoangyell@gmail.com" }
]
keywords = ["webp", "markdown", "images", "bulk", "convert", "url", "replacer", "cdn", "optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]
requires-python = ">=3.9"
dependencies = [
    "Pillow>=9.0.0",
    "requests>=2.25.0",
]

[project.scripts]
bulk-webp-url-replacer = "bulk_webp_url_replacer.cli:main"

[project.urls]
Homepage = "https://github.com/HoangGeek/bulk-webp-url-replacer"
Repository = "https://github.com/HoangGeek/bulk-webp-url-replacer"
Issues = "https://github.com/HoangGeek/bulk-webp-url-replacer/issues"
